Move run into its own package
