+++ main.go
@@ -2,9 +2,11 @@ package main
 
 import (
 	"os"
+
+	"github.com/pwr22/zoom/run"
 )
 
 func main() {
 	parseArgs()
-	os.Exit(runCmds(getCmdStrings(), *parallelism))
+	os.Exit(run.Cmds(getCmdStrings(), *parallelism))
 }
