+++ run/run_windows_test.go
@@ -1,6 +1,6 @@
 // +build !windows
 
-package main
+package run
 
 import (
 	"os/exec"
@@ -10,10 +10,9 @@ import (
 	"github.com/pwr22/zoom/job"
 )
 
-// TODO test output returned
 func TestJobRunner(t *testing.T) {
 	// we'll run two commands, one will succeed and one will be stopped to simulate error
-	job1, job2 := job.Create("sleep 0.1"), job.Create("sleep 1")
+	job1, job2 := job.Create("timeout 1"), job.Create("timeout 1")
 	jobsToRun, jobsCompleted, jobsErrored := make(chan *job.Job, 1), make(chan *job.Job, 1), make(chan *job.Job, 1)
 
 	go jobRunner(jobsToRun, jobsCompleted, jobsErrored)
@@ -42,7 +41,7 @@ func TestJobRunner(t *testing.T) {
 				t.Fatal("expected error that wait has already been called")
 			} else if err.Error() != "exec: Wait was already called" {
 				t.Fatalf("expected error that wait was already called but got: %v", err)
-			} else if _, ok := job.Err.(*exec.ExitError); !ok {
+			} else if _, ok := job.err.(*exec.ExitError); !ok {
 				t.Fatalf("expected ExitError but got %v", err)
 			}
 		}
